Add files via upload
This is a neural network ( 2-5-5-1), which can approximate any function(hopefully) to a reasonable amount of accuracy. In this particular implementation, it is approximating the Schaffer's function, which is used for testing the convergence of evolutionary computing algorithms in general. Backpropagation is used in updation of the biases and the weights.